Add the GRCon20 logos
The svg was created by inkscape (and is not flattened), so if you run
into any compatibility issues - try that first.

The png was exported from inkscape at 600 dpi, to create a
5282 x 1861, 8-bit/color RGBA, non-interlaced png.

The Car outline and gauges were done by someone else in illustrator.

The sine wave was a small GNU Plot script, exported as svg, imported
into inkscape.

The font, the "GR" is used from the "GNU Radio" logo, and the
"Con20" and "CHARLOTTE, NC" is Harbara, with a slight tweak on the "n"
to match up with the GNU Radio rings. This was all done in inkscape.

Signed-off-by: Robin Getz <robin.getz@analog.com>